// src/types.ts
export type FakerCategories =
  | 'airline'
  | 'color'
  | 'commerce'
  | 'company'
  | 'date'
  | 'finance'
  | 'git'
  | 'hacker'
  | 'image'
  | 'internet'
  | 'location'
  | 'lorem'
  | 'music'
  | 'person'
  | 'number'
  | 'phone'
  | 'science'
  | 'string'
  | 'system'
  | 'vehicle'
  | 'word'